Admin, hashing, validatins

+++ backend/services/postService.php
@@ -16,6 +16,15 @@
 
         public function createPost($post)
         {
+            if(empty($post->occasion)) {
+                throw new Exception("Occasion can't be empty.");
+            } else if(empty($post->occasionDate)) {
+                throw new Exception("Occasion date can't be empty.");
+            } else if(empty($post->location)) {
+                throw new Exception("Location can't be empty.");
+            } else if(empty($post->content)) {
+                throw new Exception("Content can't be empty.");
+            }
             $query = $this->postRepository->insertPostQuery([
                 "occasion" => $post->occasion,
                 "privacy" => $post->privacy,
@@ -72,8 +81,10 @@
         }
 
         public function answerPost($postId, $isAccepted, $userId) {
+            if(empty($postId) || empty($userId) || !isset($isAccepted)) {
+                throw new Exception("Empty postId, userId or isAccepted in answerPost.");
+            }
             $result = $this->postRepository->getAnswer($postId, $userId)["data"]->fetch(PDO::FETCH_ASSOC);
-            
             
             if(empty($result)) {
                 $this->postRepository->insertAnswer($postId, $isAccepted, $userId);
@@ -83,12 +94,18 @@
         }
 
         public function deletePost($postId) {
-        // first delete the post entries in user_post, then the post itself
+            if(empty($postId)) {
+                throw new Exception("Empty postId in deletePost.");
+            }
+            // first delete the post entries in user_post, then the post itself
             $this->postRepository->deleteAnsweredUsersPostQuery($postId);
             $this->postRepository->deletePostQuery($postId);
         }
 
         public function getIfUserAccepted($postId) {
+            if(empty($postId)) {
+                throw new Exception("Empty postId in getIfUserAccepted.");
+            }
             $result =  $this->postRepository->getAnswer($postId, $_SESSION['userId'])["data"]->fetch(PDO::FETCH_ASSOC);
 
             if (empty($result)) {
